fix(UIKIT-327): вынесение логики в кастомный хук
