+++ packages/ui/src/OverflowTypography/hooks/useOverflowed/index.ts
@@ -1,0 +1,1 @@
+export * from './useOverflowed';
